requirements.txt
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
english-dictionary==1.0.24
evdev==1.5.0
idna==3.3
lxml==4.9.1
MouseInfo==0.1.3
numpy==1.23.0
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytesseract==0.3.9
python-xlib==0.31
python3-xlib==0.15
pytweening==1.0.4
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.9
